requirements.txt
pytorch>=1.7.0